backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.40.0
web3==6.20.3
requests==2.32.3
tree-sitter==0.23.2
tree-sitter-solidity==1.2.11
